+++ CHANGELOG.md
@@ -4,3 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.2] - 2024-12-17
+### Added
+- Added [WooCommerce HPOS](https://woocommerce.com/document/high-performance-order-storage/) compatibility declaration
+### Fixed
+- Changed plugin title to Nordea Finance Erämaksu
+- Made Nordea Finance logo smaller
+- Moved legal text to be predifined and constant